feedback_loop/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.5.0
